File IO Move, minor change to Generator
File IO now has its own c++ and header file. Slight change to how the
processes are generated. Moved the PID to be only generated once the
process is being read in.
